requirements.txt
# Finance Agent Python Dependencies

# Core dependencies
numpy<2.0.0  # 锁定 numpy 版本以避免与 torch 的冲突

# Database
aiosqlite>=0.19.0

# ChromaDB Vector Database
chromadb>=0.4.0
sentence-transformers>=3.0.0  # 启用 sentence-transformers，使用兼容版本

# Claude Agent SDK (官方 SDK)
claude-agent-sdk>=0.1.0

# Anthropic Claude SDK (备用)
anthropic>=0.18.0

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
websockets>=12.0

# HTTP Client
httpx>=0.26.0
aiohttp>=3.9.0  # 添加 aiohttp（测试脚本需要）

# Python dotenv for environment variables
python-dotenv>=1.0.0

# WebSocket client (for testing)
python-socks>=2.0.0

# File watching for hot reload
watchdog>=3.0.0